refactor: move waitForLatestProcessVersion into zeebeClient utils
Uses the existing Camunda REST client (searchProcessDefinitions)
instead of a hand-rolled fetch with manual auth, and keeps the spec
file free of helper plumbing.
